// src/types/Monitoring.types.d.ts
export interface thresholdType {
  vibration?: {
    red?: number;
    yellow?: number;
    green?: number;
  };
  noise?: {
    red?: number;
    yellow?: number;
    green?: number;
  };
  temperature?: {
    red?: number;
    yellow?: number;
    green?: number;
  };
}
export interface telemetryDataPointsType {
  temperature: number;
  vibration: number;
  noise: number;
  date: number;
}

export interface deviceDataType {
  temperature: number;
  vibration: number;
  noise: number;
  date: string;
  status : string
}
// ...
export interface realTimeDataType {
  goodValue?: number;
  averageValue?: number;
  peakValue?: number;
  xAxis?: string;
  presentValue?: number;
}